configurable-http-proxy >= 4
Recent conda-forge builds of configurable-http-proxy are tied to a node version. If a new node version without a corresponding configurable-http-proxy build is installed the resolver may pick an ancient version of configurable-http-proxy that doesn't specify the require node version. Adding a minimum here should mean the node version is downgraded to one for which a configurable-http-proxy build is available.